fix(totd): load only when there is tips content for locale
